requirements.txt
napalm
ncclient
pyyaml
xmltodict